refactor: move Solana publishing to a separate module
